requirement.txt
flask
ultralytics
torch
opencv-python-headless
Pillow
